Changes 1/28
I worked for a few extra hours and was able to get rid of most of the errors, including segmentation error. The problem that still exist is that the array doens't seem to be updating, despite some of the values changing. If you need any parts to be clarified, just send me a message - email would be guaranteed.